requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-ajax-datatable @ git+https://github.com/morlandi/django-ajax-datatable.git@ac0f4edde748674dc9b960b9a4876bee39766132
django-debug-toolbar==5.2.0
django-modeltranslation==0.19.14
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pygame==2.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
translation==1.0.5
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==22.0.0